+++ src/Entry/Handler.tsx
@@ -3,9 +3,6 @@ require("fetch-everywhere");
 
 process.env.MODE = 'server';
 
-import * as cluster from 'cluster';
-import * as os from 'os';
-
 let scope: any = process.env.HANDLER_SCOPE || 'Main';
 
 import { getConfig, readConfig } from '../Modules/Config';
@@ -16,30 +13,9 @@ Log.init();
 
 import * as ORM from '../Modules/ORM';
 import * as Handler from '../Modules/Handler';
+import { runCluster } from '../Server/Lib/EntryRunner';
 
-export const run = () => {
-  if (!getConfig().cores) {
-    const cronManager = new Handler.JobManager(scope);
-    cronManager.init();
-  } else {
-    if (cluster.isMaster) {
-      const numCPUs = getConfig().cores == 'auto' ? os.cpus().length : getConfig().cores;
-
-      for (let i = 0; i < numCPUs; i++) {
-        cluster.fork();
-      }
-
-      cluster.on('exit', function(worker, code, signal) {
-        Log.logWarn('Worker ' + worker.process.pid + ' died.');
-        cluster.fork();
-      });
-
-      cluster.on('listening', function(worker, address) {
-        Log.logInfo('Worker started with PID ' + worker.process.pid + '.');
-      });
-    } else {
-      const cronManager = new Handler.JobManager(scope);
-      cronManager.init();
-    }
-  }
-}
+export const run = () => runCluster(() => {
+  const cronManager = new Handler.JobManager(scope);
+  cronManager.init();
+});
