+++ src/Entry/Worker.tsx
@@ -16,32 +16,9 @@ import * as Log from '../Modules/Log';
 Log.init();
 
 import * as Worker from '../Modules/Worker';
-import * as Translation from '../Modules/Translation';
+import { runCluster } from '../Server/Lib/EntryRunner';
 
-export const run = () => {
-  if (!getConfig().cores) {
-    const cronManager = new Worker.CronManager(scope);
-    cronManager.init();
-  } else {
-    if (cluster.isMaster) {
-      const numCPUs = getConfig().cores == 'auto' ? os.cpus().length : getConfig().cores;
-
-      for (let i = 0; i < numCPUs; i++) {
-        cluster.fork();
-      }
-
-      cluster.on('exit', function(worker, code, signal) {
-        Log.logWarn('Worker ' + worker.process.pid + ' died.');
-        cluster.fork();
-      });
-
-      cluster.on('listening', function(worker, address) {
-        Log.logInfo('Worker started with PID ' + worker.process.pid + '.');
-      });
-
-    } else {
-      const cronManager = new Worker.CronManager(scope);
-      cronManager.init();
-    }
-  }
-}
+export const run = () => runCluster(() => {
+  const cronManager = new Worker.CronManager(scope);
+  cronManager.init();
+});
